+++ MkDocs/toplevel/docs/release_notes.md
@@ -18,6 +18,7 @@
 - Jignesh Prajapati - <jignesh.prajapati@ibm.com>
 - Radhika Surani - <Radhika.Surani@ibm.com>
 - Mayur Sojitra - <mayur.sojitra@ibm.com>
+- Margi Shah - <margi.shah1@ibm.com>
  
 
 ---
@@ -26,6 +27,7 @@
 
 |Date      |By             | Description                                                                            |
 |:---------|:--------------|:---------------------------------------------------------------------------------------|
+|2025-07-23|Margi Shah     |Added the initial version of [Maximo Monitor Managed Hierarchy lab](../monitor_managed_hierarchy_9.1/) for MAS 9.1. |
 |2025-07-01|Rishi Kant    |Added the initial version of [Monitor Cisco Webex Integration Lab](../monitor_cisco_webex_integration_9.1/) for MAS 9.1. |
 |2025-07-01|Kajal Lalwani    |Added the initial version of [ Monitor Maximo Real Estate and Facilities Lab](../monitor_MREF_9.1/) for MAS 9.1. |
 |2025-07-04|Utsav Patel    |Added the initial version of [Maximo Monitor Alerts lab](../monitor_managed_alerts_9.1/) for MAS 9.1. |